Fix problems with writing RIFF
The version was flipped around, causing devenv not being able to understand. Now devenv and the app understand palettes that have been written out, even if they've been modified.

Now, devenv (and likely other programs) may choke on some colours and fall back to monochrome hatches, but I suppose this is an issue with them.