Actually get DPI awareness (very basic)
Not the best, but don't feel ready to force 4.7 yet, and a lot of the methods like manifests don't seem to work - you call this instead. FWIW, linking WPF in seems to also call this, but we don't need to do that...

FIXME: Add a manifest so we can look for Windows 8.1, not 8.0.